Add Group/PermissionSettings controller with test
